Create random_walk_test.r
This script compares animal movement trajectories to a set of randomly generated trajectories (generated by randomizing the true trajectories' steps' angles and distances. It compares the true trajectories' R-squared and linearity value to the distribution of R-squared and linearity values of the random replicates, following a procedure implemented in the Site Fidelity tool of Animal Movement Extension for ArcView (http://alaska.usgs.gov/science/biology/spatial/gistools/animalmovementdoca.pdf).

It takes inputs of animal locations and a shapefile with one polygon which represents the area available to the animal (this constrains the random walks). Plots are output displaying the trajectory and random walks, and a output table with r-squared, linearity, and p-values for each track.